Trait parables_testing::abi::Constructor

pub trait Constructor {
    const ITEM: &'static str;
    const BIN: &'static str;
    const SOURCE_MAP: Option<&'static str>;
    const RUNTIME_BIN: Option<&'static str>;
    const RUNTIME_SOURCE_MAP: Option<&'static str>;
}

Associated Constants

const ITEM: &'static str

Name of the constructor item, used for linking.

const BIN: &'static str

Access the code to deploy for this constructor.

const SOURCE_MAP: Option<&'static str>

Access the source map for the type this constructor is associated with.

const RUNTIME_BIN: Option<&'static str>

Access the runtime code being deployed.

const RUNTIME_SOURCE_MAP: Option<&'static str>

Access the runtime source map for the type this constructor is associated with.