Struct parables_testing::evm::Session

pub struct Session<'a, 'b: 'a> { /* fields omitted */ }

Methods

impl<'a, 'b: 'a> Session<'a, 'b>

pub fn deploy<F>(&mut self, f: F, call: Call) -> Result<Address, CallError> where
F: ContractFunction<Output = Address>,

Deploy the contract with the given code.

pub fn deploy_code(
&mut self,
code: Vec<u8>,
call: Call
) -> Result<Address, CallError>

Deploy the contract with the given code.

pub fn call<F>(
&mut self,
address: Address,
f: F,
call: Call
) -> Result<F::Output, CallError> where
F: ContractFunction,

Perform a call against the given contract function.

pub fn fallback(
&mut self,
address: Address,
call: Call
) -> Result<(), CallError>

Perform a call against the given contracts' fallback function.

pub fn logs(&self) -> &HashMap<Hash, Vec<LogEntry>>

Access all logs.

pub fn has_logs(&self) -> bool

Check if we still have unclaimed logs.

pub fn drain_logs<P>(&mut self, filter: Filter<P>) -> Result<Vec<P::Log>, Error> where
P: ParseLog,

Drain logs matching the given filter that has been registered so far.

pub fn drain_logs_with_sender<P>(
&mut self,
filter: Filter<P>
) -> Result<Vec<(Address, P::Log)>, Error> where
P: ParseLog,

Drain logs matching the given filter that has been registered so far.

Include who sent the logs in the result.