Struct parables_testing::ledger::Ledger

pub struct Ledger<'a, S> where
    S: LedgerState,  { /* fields omitted */ }

Methods

impl<'a> Ledger<'a, ()>

pub fn empty(evm: &'a Evm) -> Ledger<'a, ()>

Construct a new empty ledger that doesn't have any specialized state.

impl<'a, S> Ledger<'a, S> where
S: LedgerState,

pub fn new(evm: &'a Evm, state: S) -> Ledger<S>

Construct a new ledger.

To construct a ledger without state, use Ledger::empty().

pub fn sync(&mut self, address: Address) -> Result<(), Error>

Synchronize the ledger against the current state of the virtual machine.

pub fn verify(self) -> Result<(), Error>

Go through each registered account, and verify their invariants.

pub fn add<V>(&mut self, address: Address, value: V) where
V: Into<U256>,

Add to the balance for the given address.

pub fn sub<V>(&mut self, address: Address, value: V) where
V: Into<U256>,

Subtract from the balance for the given address.

pub fn state(&mut self, address: Address) -> &mut S::Instance

Access the mutable state for the given address.