Struct parables_testing::prelude::ethabi::Contract

pub struct Contract {
    pub constructor: Option<Constructor>,
    pub functions: HashMap<String, Function, RandomState>,
    pub events: HashMap<String, Event, RandomState>,
    pub fallback: bool,
}

API building calls to contracts ABI.

Fields
constructor: Option<Constructor>

Contract constructor.

functions: HashMap<String, Function, RandomState>

Contract functions.

events: HashMap<String, Event, RandomState>

Contract events.

fallback: bool

Contract has fallback function.

Methods

impl Contract

pub fn load<T>(reader: T) -> Result<Contract, Error> where
T: Read,

Loads contract from json.

pub fn constructor(&self) -> Option<&Constructor>

Creates constructor call builder.

pub fn function(&self, name: &str) -> Result<&Function, Error>

Creates function call builder.

pub fn event(&self, name: &str) -> Result<&Event, Error>

Creates event decoder.

Important traits for Functions<'a>
impl<'a> Iterator for Functions<'a> type Item = &'a Function;

pub fn functions(&self) -> Functions

Iterate over all functions of the contract in arbitrary order.

Important traits for Events<'a>
impl<'a> Iterator for Events<'a> type Item = &'a Event;

pub fn events(&self) -> Events

Iterate over all events of the contract in arbitrary order.

pub fn fallback(&self) -> bool

Returns true if contract has fallback