Struct parables_testing::test_runner::TestRunner

pub struct TestRunner<'a> { /* fields omitted */ }

A scaffolding that runs tests very efficiently.

Methods

impl<'a> TestRunner<'a>

pub fn new() -> Self

Build a new test runner.

pub fn module<'m>(
&'m mut self,
name: impl Into<Cow<'a, str>>
) -> ModuleRunner<'m, 'a>

Create a module runner.

pub fn run(self, reporter: &Reporter<'a>) -> Result<(), Error>

Run by reading filters from argv.

pub fn run_with_filters<F>(
self,
filters: F,
reporter: &Reporter<'a>
) -> Result<(), Error> where
F: IntoIterator<Item = String>,

Run all registered tests, while applying the given filter on their name.

All strings specified in the filter must be apart of the name of a test to include it.

Note: this installs a panic hook, so mixing this with another component that fiddles with the hook will cause unexpected results.